feat: ticket creation updated (#603)
* ticket creation V2 to API

* prettier write

* fix double slash problem

* fix tests

* Remove automatic v1 prefixing

Make sure that v1 prefix is explicitly set if it is needed, this way
api.qminder.com/ticket can exclude the v1 prefix.

* fix: only one copy of each API endpoint

We decided to ensure that we don't keep multiple versions of a single
API available in the JS SDK. There are other options, like staying at
an older version, or calling the REST API manually.

* update docs + don't use template string if not required

* do not export BaseInputFieldRequest

* input-field-request

---------

Co-authored-by: Siim Raud <siim@qminder.com>